ucm-imx8m-mini: Introduce fdt_file autoselection
In order to support the PCB revision 1.4, equipped with the WiFi module EL2 by
Murata, instead of WB5+ by Laird, introdused the autoselection of fdt file to be
loaded. The selection is based on the presence of the sub-string 'WBM' in the
product option line (M for Murata).
A non-empty string, assigned to the u-boot environment variable 'fdt_file'
overrides the autoselection.

Signed-off-by: Kirill Kapranov <kirill.kapranov@compulab.co.il>